+++ index.html
@@ -52,6 +52,6 @@
                 
     </main>
     
-    <script type="module" src="./src/app.js"></script>
+    <script src="./src/main.js"></script>
 </body>
 </html>
