--- src/app.js
@@ -1,9 +1,0 @@
-import { displayPokemon } from "./modules/ui.js";
-
-// let saveNotes = JSON.parse(localStorage.getItem("note")) || [];
-// let favorites = localStorage.getItem("favorites");
-// favorites = favorites ? JSON.parse(favorites) : [];
-
-displayPokemon();
-
-
