+++ src/journal.html
@@ -31,13 +31,13 @@
 
         <div id="hero-image" class="mt-40 mx-[20%]">
             <!--only if we want one-->
-            <img src="./img/Pokemon.png" class="rounded-xl h-auto max-w-[300px]" alt="Description">
+            <img src="img/Pokemon.png" class="rounded-xl h-auto max-w-[300px]" alt="Description">
         </div>
     </section>
     </header>
 
     <main id="cards-container" class="w-2/3 grid grid-cols-3 auto-rows-min gap-2 overflow-y-auto scrollbar" ></main>
     
-    <script type="module" src="./journal.js"></script>
+    <script src="journal.js"></script>
 </body>
 </html>
